Rename the core package to gagavatar and make it installable
"core" is not an importable identity outside this repository and
collides with other projects' package names. Rename the package to
gagavatar, update the imports and the GAGAvatar_track submodule path,
and add minimal pyproject metadata so the repository can be installed
with pip (including editable installs) and imported as `gagavatar` by
external integrations. Pure rename otherwise: no content changes.
