+++ gagavatar/libs/utils_perceptual.py
