+++ pyproject.toml
@@ -1,0 +1,14 @@
+[build-system]
+requires = ["setuptools>=68"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "gagavatar"
+version = "0.1.0"
+description = "GAGAvatar: Generalizable and Animatable Gaussian Head Avatar"
+readme = "README.md"
+requires-python = ">=3.10"
+
+[tool.setuptools.packages.find]
+include = ["gagavatar*"]
+namespaces = true
